feat: make the bot able to respond intelligently to the pie rule and improve gui related to this mechanic
