+++ gamey/src/bot/smart_bot.rs
@@ -5,7 +5,8 @@
 
 use crate::{Coordinates, GameY, YBot};
 
-use super::minimax::choose_move_with_minimax;
+use super::minimax::{choose_move_with_minimax, choose_pie_opening_with_minimax, decide_pie_with_minimax};
+use super::ybot::PieChoice;
 
 /// A bot that runs minimax with a 1 000–3 000 ms time budget.
 ///
@@ -19,6 +20,14 @@ impl YBot for SmartBot {
 
     fn choose_move(&self, game: &GameY) -> Option<Coordinates> {
         choose_move_with_minimax(game, 1000, 3000)
+    }
+
+    fn choose_pie_opening(&self, game: &GameY) -> Option<Coordinates> {
+        choose_pie_opening_with_minimax(game, 2000)
+    }
+
+    fn decide_pie(&self, game: &GameY) -> PieChoice {
+        decide_pie_with_minimax(game, 2000)
     }
 }
 
@@ -37,5 +46,28 @@ mod tests {
         let coords = SmartBot.choose_move(&game);
         assert!(coords.is_some());
         assert!(coords.unwrap().is_valid(3));
+    }
+
+    #[test]
+    fn test_smart_bot_pie_opening_returns_valid_move() {
+        let game = GameY::new(3);
+        let coords = SmartBot.choose_pie_opening(&game);
+        assert!(coords.is_some());
+        assert!(coords.unwrap().is_valid(3));
+    }
+
+    #[test]
+    fn test_smart_bot_decide_pie_returns_a_choice() {
+        use crate::{Coordinates, Movement, PlayerId};
+
+        let mut game = GameY::new(5);
+        game.add_move(Movement::Placement {
+            player: PlayerId::new(0),
+            coords: Coordinates::new(2, 1, 1),
+        })
+        .unwrap();
+
+        let choice = SmartBot.decide_pie(&game);
+        assert!(choice == PieChoice::Keep || choice == PieChoice::Swap);
     }
 }
