+++ webapp/src/controllers/useGameYController.ts
@@ -1,4 +1,6 @@
 import { useState, useEffect, useCallback, useRef } from 'react'
+
+const SWAP_ANIM_MS = 600
 import { useParams, useNavigate } from 'react-router-dom'
 import type { GameState, ChatMessage, Move, PieDecision, PlayerColor, TimerState } from '@/types'
 import { useAuth } from '@/contexts/AuthContext'
@@ -18,6 +20,16 @@ export function useGameYController() {
     const [isBotThinking, setIsBotThinking] = useState(false)
     const [isPieDecisionLoading, setIsPieDecisionLoading] = useState(false)
     const [isSwapAnimating, setIsSwapAnimating] = useState(false)
+    // Set immediately when the human commits to swap so the highlight can show
+    // the stone as red before the API responds (pre-coloring).
+    const [swapCommitted, setSwapCommitted] = useState(false)
+    // True while the server is auto-resolving the bot's pie decision (bot is P2).
+    // Used to show the panel + spinner during the playMove round-trip.
+    const [isBotResolvingPie, setIsBotResolvingPie] = useState(false)
+    // When the bot auto-swaps we animate the contested stone. We track its
+    // position separately because after the API response lastMove points to
+    // the bot's follow-up move, not the swapped stone.
+    const [swapAnimationStone, setSwapAnimationStone] = useState<{ row: number; col: number } | null>(null)
 
     const [liveTimer, setLiveTimer] = useState<TimerState | null>(null)
     const clockStartedAtRef = useRef(Date.now())
@@ -164,10 +176,33 @@ export function useGameYController() {
             clockStartedAtRef.current = Date.now()
         }
 
+        // Detect if the server will auto-resolve a bot pie decision this round-trip:
+        // pie rule enabled + this is the very first move + bot is the deciding player (P2).
+        const willBotResolvePie =
+            game.config.pieRule === true &&
+            game.moves.length === 0 &&
+            game.config.mode === 'pve' &&
+            game.players.player2.isBot === true
+
         setIsBotThinking(game.config.mode === 'pve')
+        if (willBotResolvePie) setIsBotResolvingPie(true)
         try {
             const updated = await gameService.playMove(gameId, row, col, game.currentTurn as PlayerColor, effectiveToken)
             setGame(updated)
+
+            // If the bot was resolving pie and the contested stone is now owned
+            // by player2 (the bot swapped), play the swap animation on that cell.
+            if (willBotResolvePie) {
+                const contestedCell = updated.board[row]?.[col]
+                if (contestedCell?.owner === 'player2') {
+                    setSwapAnimationStone({ row, col })
+                    setIsSwapAnimating(true)
+                    setTimeout(() => {
+                        setIsSwapAnimating(false)
+                        setSwapAnimationStone(null)
+                    }, SWAP_ANIM_MS)
+                }
+            }
         } catch (err) {
             setGame(snapshot)
             const msg = err instanceof Error ? err.message : 'Failed to play move'
@@ -175,6 +210,7 @@ export function useGameYController() {
             console.error('Failed to play move:', err)
         } finally {
             setIsBotThinking(false)
+            setIsBotResolvingPie(false)
         }
     }, [game, gameId, canPlay, effectiveToken, liveTimer])
 
@@ -209,7 +245,14 @@ export function useGameYController() {
     const handlePieDecision = useCallback(async (decision: PieDecision) => {
         if (!game || !gameId) return
         setIsPieDecisionLoading(true)
-        if (decision === 'swap') setIsSwapAnimating(true)
+
+        // For swap: start the animation immediately and guarantee it runs for at
+        // least SWAP_ANIM_MS regardless of how fast the API responds.
+        let swapAnimTimer: ReturnType<typeof setTimeout> | null = null
+        if (decision === 'swap') {
+            setIsSwapAnimating(true)
+            swapAnimTimer = setTimeout(() => setIsSwapAnimating(false), SWAP_ANIM_MS)
+        }
 
         // Optimistic update for swap: immediately change the stone from Blue (player1)
         // to Red (player2). The 700 ms fill transition in GameYCell makes this visible.
@@ -228,15 +271,19 @@ export function useGameYController() {
             }
         }
 
+            if (decision === 'swap') setSwapCommitted(true)
         try {
             const updated = await gameService.decidePie(gameId, decision, effectiveToken)
             setGame(updated)
         } catch (err) {
             setGame(snapshot) // revert optimistic update on error
+            if (swapAnimTimer) clearTimeout(swapAnimTimer)
+            setIsSwapAnimating(false)
+            setSwapCommitted(false)
             console.error('Failed to submit pie decision:', err)
         } finally {
             setIsPieDecisionLoading(false)
-            setTimeout(() => setIsSwapAnimating(false), 750)
+            setSwapCommitted(false)
         }
     }, [game, gameId, effectiveToken])
 
@@ -258,6 +305,9 @@ export function useGameYController() {
         isBotDecidingPie,
         isPieDecisionLoading,
         isSwapAnimating,
+        isBotResolvingPie,
+        swapAnimationStone,
+        swapCommitted,
         canPlay: canPlay(),
         handleCellClick,
         handlePieDecision,
